// src/app/interfaces/checkmanip.ts
export interface CheckManip{
    id?: number;	
    folio: number;	
    num_empleado:number;
    fecha:string;
    manipulador: string;
    codigo_parada1: string;	
    p_hora_inicio1: string;	
    p_hora_termino1: string;	
    p_litros1: string;	
    p_observaciones1: string;	
    codigo_parada2: string;	
    p_hora_inicio2: string;	
    p_hora_termino2: string;	
    p_litros2: string;	
    p_observaciones2: string;	
    verificar_torreta: string;	
    verificar_claxon: string;	
    freno_estacionamiento: string;	
    freno_servicio: string;	
    freno_parking: string;	
    alarma_reversa: string;	
    estado_asiento_cinturon: string;	
    funcionamiento_estabilizadores: string;	
    retraccion_extension_pluma: string;	
    montaje_canastilla_cuchillas: string;	
    luces_faros: string;	
    extintor_manual: string;
    articulacion_ruedas: string;	
    panel_control: string;	
    nivel_combustible: string;	
    nivel_aceite_motor: string;	
    nivel_aceite_hidraulico: string;	
    nivel_aceite_transmision: string;	
    nivel_anticongelante: string;	
    nivel_mandos_finales: string;	
    inspeccionar_golpes_equipo: string;	
    indicador_serv_filtros_aire: string;	
    sep_agua_sistema_combustible: string;	
    tablero_instrumentos: string;	
    revisar_engrasar_crucetas: string;	
    fugas_aceite: string;	
    terminales_direccion: string;	
    presion_llantas: string;	
    limpieza_equipo: string;	
    bisagra_cierre_acceso: string;	
    seguro_tope_bisagra: string;	
    regilla_piso: string;	
    fisuras_estructuras_soldadura: string;	
    tuercas_abrazaderas: string;	
    observaciones: string;	
    nombre_operador: string;	
    nombre_supervisor: string;	
    created_at?: string;	
    updated_at?: string;	
}